+++ components/calendar/Constants.js
@@ -1,0 +1,3 @@
+export default {
+  PREFIX_CLS: 'ant-calendar',
+};
