+++ style/components/calendar.less
@@ -1,105 +1,3 @@
-div.rc-calendar-full { width: 280px; }
-div.rc-calendar-fullscreen { width: 100%; }
-
-.rc-calendar-fullscreen {
-  .rc-calendar-table {
-    table-layout: fixed;
-  }
-  
-  .rc-calendar-month-cell-notes {
-    text-align: center;
-    color: #666;
-
-    > section {
-      font-size: 24px;
-    }
-  }
-
-}
-
-.calendar-notes {
-  height: 9px;
-  line-height: 8px;
-  text-align: center;
-  overflow: hidden;
-  border-radius: 4px;
-  bottom: -5px;
-  left: 4px;
-  font-size: 8px;
-  margin: auto;
-  width: 19px;
-  cursor: pointer;
-
-  &-overflow {
-    border: 1px solid #d9d9d9;
-  }
-
-  span&-node-gray {
-    color: #ccc;
-  }
-  span&-node-warn {
-    color: #fac450;
-  }
-  span&-node-normal {
-    color: #2db7f5;
-  }
-  span&-node-error {
-    color: #f60;
-  }
-
-  &-date {
-    width: 20px; 
-    height: 20px;
-    line-height: 20px;
-  }
-}
-
-span.calendar-notes-date-full {
-  overflow: auto;
-  padding: 0 8px 8px 8px;
-}
-
-.calendar-notes-list {
-  list-style: none;
-  text-align: left;
-  margin: 0;
-  padding: 0;
- 
-
-  & * { vertical-align: middle; }
-
-  > li {
-    color: #999;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-
-    > span:first-child {
-    font-size: 9px;
-    margin-right: 4px;
-  } 
-  }
-
-  &-node-gray {
-    color: #ccc;
-  }
-  &-node-warn {
-    color: #fac450;
-  }
-  &-node-normal {
-    color: #2db7f5;
-  }
-  &-node-error {
-    color: #f60;
-  }
-}
-
-.rc-tooltip {
-  .rc-tooltip-inner {
-    padding: 8px 12px;
-  }
-  
-}
-
-
-
+@import "calendar/index";
+@import "calendar/NoteList";
+@import "calendar/Notes";
