+++ style/components/calendar/NoteList.less
@@ -1,0 +1,36 @@
+.@{calendar-prefix-cls}-notes {
+  &-list {
+    list-style: none;
+    text-align: left;
+    margin: 0;
+    padding: 0;
+   
+
+    & * { vertical-align: middle; }
+
+    > li {
+      color: #999;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+
+      > span:first-child {
+      font-size: 9px;
+      margin-right: 4px;
+    } 
+    }
+
+    &-node-gray {
+      color: #ccc;
+    }
+    &-node-warn {
+      color: #fac450;
+    }
+    &-node-normal {
+      color: #2db7f5;
+    }
+    &-node-error {
+      color: #f60;
+    }
+  }
+}
